% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{create_instance}
\alias{create_instance}
\title{Create instance object}
\usage{
create_instance(of, with = NA, inherits = TRUE, attributes = NULL)
}
\arguments{
\item{of}{\code{\link{character}}.
Name of the class for which an instance should be created.}

\item{with}{\code{\link{ANY}}.
Any object that should act as the actual data of the instance.}

\item{inherits}{\code{\link{logical}}.
\code{TRUE}: resulting object inherits from class(es) of \code{with}.
\code{FALSE} (default): resulting object only has class \code{of}.}

\item{attributes}{\code{\link{list}}
Attributes to be passed along to the instance. Supply as a named list.}
}
\description{
Convenience wrapper around `structure()`.
}
\details{
None yet.
}
\examples{

# Intro -------------------------------------------------------------------

create_instance(of = "test_class")
create_instance(of = "test_class", with = letters[1:3])
create_instance(of = "test_class", with = letters[1:3], inherits = FALSE)

# Additional attributes ---------------------------------------------------

create_instance(of = "test_class", with = letters[1:3],
  attributes = list(attr_1 = TRUE, attr_2 = "yep!"))
create_instance(of = "test_class", with = letters[1:3],
  attributes = list(my_attrs = list(attr_1 = TRUE, attr_2 = "yep!")))

# Preserve data frames ----------------------------------------------------

# A data frame is automatically "reduced" to its list structure if no
# inheritance is involved
structure(mtcars, class = "test_class")

# Therefore it seems to be best to throw an error in case the user wants to try
# this in order to make it explicit to chose a list input if that's what he
# wants
create_instance(of = "test_class", with = mtcars, inherits = FALSE)
create_instance(of = "test_class", with = unclass(mtcars), inherits = FALSE)
}
\references{
\url{http://github.com/rappster/instancer}
}
\seealso{
\code{
  	\link[instancer]{TODO}
}
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
