% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{createInstance}
\alias{createInstance}
\title{Create instance object}
\usage{
createInstance(of, with = NA)
}
\arguments{
\item{of}{\code{\link{character}}.
Name of the class for which an instance should be created.}

\item{with}{\code{\link{ANY}}.
Any object that should act as the actual data of the instance.}
}
\description{
Convenience wrapper around `structure()`.
}
\details{
None yet.
}
\examples{
createInstance(of = "Testclass")
createInstance(of = "Testclass", with = letters[1:3])
}
\author{
Janko Thyson \email{janko.thyson@gmail.com}
}
\references{
\url{http://github.com/rappster/instancer}
}
\seealso{
\code{
  	\link[instancer]{TODO}
}
}

